requirements.txt
Flask~=1.0.3
pandas~=1.2.3
Cython==0.29.24
gunicorn==20.1.0
python-telegram-bot==13.7
requests~=2.25.0
pyzbar~=0.1.8
opencv-python~=4.5.4.58